update contract verifier in each module
