+++ packages/modules/uniswap-v3/src/utils/index.ts
@@ -1,2 +1,1 @@
 export * from '@utils/contract-verifier'
-export * from '@utils/irm'
