+++ packages/modules/uniswap-v3/tsconfig.json
@@ -20,6 +20,7 @@
       "@actions/*": ["./src/actions/*"],
       "@contracts/*": ["./src/contracts/*"],
       "@scripts/*": ["./src/scripts/*"],
+      "@utils/*": ["./src/utils/*"],
       "@/*": ["*"]
     }
   },
