+++ packages/modules/uniswap-v3/tsup.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'tsup'
 
 export default defineConfig({
   name: 'uniswap-v3',
-  entry: ['src/index.ts', 'src/actions/index.ts'],
+  entry: ['src/index.ts', 'src/actions/index.ts', 'src/utils/index.ts'], // Entry points
   format: ['esm', 'cjs'], // Build for ESmodules and Commonjs
   dts: true, // Generate declaration file (.d.ts)
   sourcemap: false,
